requirements.txt
google-analytics-data==0.17.0
pandas==2.0.3
google-auth-oauthlib==1.0.0
python-dotenv==1.0.0